ML/feature_3/requirements.txt
datasets
transformers
torch
pyyaml
sentence-transformers
scikit-learn
pandas
